requirements.txt
Django==4.0.3
django_environ==0.8.1
djangorestframework
psycopg2
django-filter
celery